refactor: PostCardActionsのリアクションロジックをuseReactionsフックに抽出
リアクション状態管理・取得・トグルロジック（約30行）を
hooks/useReactions.tsに移動し、PostCardActionsをUI専念に

closes #1497
