+++ frontend/src/hooks/index.ts
@@ -51,3 +51,4 @@ export { useConfirm } from './useConfirm';
 export { useOnboarding } from './useOnboarding';
 export { useYoutube } from './useYoutube';
 export { useCommentLike } from './useCommentLike';
+export { useReactions } from './useReactions';
